Reorganise data entry validation code
